+++ CHANGELOG.md
@@ -40,7 +40,7 @@
 - Enh #403, #404: Use `DbArrayHelper::arrange()` instead of `DbArrayHelper::index()` method (@Tigrov)
 - New #397: Realize `Schema::loadResultColumn()` method (@Tigrov)
 - New #407: Use `DateTimeColumn` class for datetime column types (@Tigrov)
-- New #408: Implement `DMLQueryBuilder::upsertWithReturningPks()` method (@Tigrov)
+- New #408, #410: Implement `DMLQueryBuilder::upsertReturning()` method (@Tigrov)
 
 ## 1.3.0 March 21, 2024